Fixing issues found in the QA sets (#10)
* fixing typo in test json

* testing changes

* testing

* testing

* not an issue

* adding more checks

* enab.ing

* testing

* typo

* enabling

* dont delete certain stuff

* testing

* forgot self argument

* tying to dump diffexp object file

* making diffexp folder correctly

* testing

* trying to fix

* removing experimental stuff

* removing sam file from outputs

* removing diffexp import for now

* enable fastqc

* making more robust

+++ service-scripts/App-RNASeq.pl
@@ -204,14 +204,30 @@ sub process_rnaseq {
     my $time2 = `date`;
     my $outdir = "$tmpdir";
     
-    # Remove genome directory
-    my $ref_id   = $params->{reference_genome_id};
-    my $ref_dir  = "$tmpdir/$ref_id";
-    rmtree([ "$ref_dir" ]);
+    if ($called_localize_params) {
+        # Remove genome directory
+        my $ref_id   = $params->{reference_genome_id};
+        my $ref_dir  = "$tmpdir/$ref_id";
+        if ( -d $ref_dir ) {
+            rmtree([ "$ref_dir" ]);
+        }
 
-    # Remove TPMCalculator directory
-    my $tpm_dir = "$tmpdir/TPMCalculator";
-    rmtree([ "$tpm_dir" ]);
+        # Remove TPMCalculator directory
+        my $tpm_dir = "$tmpdir/TPMCalculator";
+        if ( -d $tpm_dir ) {
+            rmtree([ "$tpm_dir" ]);
+        }
+    }
+
+    # save diffexp stuff
+    #my $diffexp_name = "diff_exp";
+    #my $diffexp_folder = "$output/.$diffexp_name";
+    #my $diffexp_file = "$output/$diffexp_name";
+
+    # copy over diffexp stuff
+    #if ( -e $diffexp_folder ) {
+    #    $app->workspace->save_file_to_file("$diffexp_folder", {},"$output_folder/.$diffexp_name", "job_result", 1);
+    #}
 
     save_output_files($app,$outdir);
     write_output("Start: $time1"."End:   $time2", "$tmpdir/DONE");
@@ -260,9 +276,8 @@ sub run_bvbrc_rnaseq {
     my $ref_id   = $params->{reference_genome_id} or die "Reference genome is required\n";
     my $output_name = $params->{output_file} or die "Output name is required\n";
     my $host_ftp = defined($params->{host_ftp}) ? $params->{host_ftp} : undef;
-    print STDERR "here\n";
-    my $diffexp_name = defined($params->{diffexp_name}) ? $params->{diffexp_name} : "diff_exp";
-    my $diffexp_folder = "$outdir/$diffexp_name";
+    my $diffexp_name = "diff_exp";
+    my $diffexp_folder = "$outdir/.$diffexp_name";
     my $diffexp_file = "$outdir/$diffexp_name";
     my $ref_dir  = prepare_ref_data_rocket($ref_id, $tmpdir, $host, $host_ftp);
     #my $unit_test = defined($params->{unit_test}) ? $params->{unit_test} : undef;
@@ -280,7 +295,7 @@ sub run_bvbrc_rnaseq {
     push @cmd, ("-p", $pstring);
     push @cmd, ("-o", $outdir);
     push @cmd, ("-g", $ref_dir);
-    push @cmd, ("-d", $diffexp_name);
+    push @cmd, ("-d", $diffexp_folder);
     push @cmd, ("--jfile", $jdesc);
     push @cmd, ("--sstring", $sstring);
     #if ($unit_test) {
@@ -355,7 +370,7 @@ sub run_bvbrc_rnaseq {
 	push(@outputs, [$txt, 'txt']);
     }
     
-    push @outputs, [ "$outdir/$ref_id/gene_exp.gmx", 'diffexp_input_data' ] if -s "$outdir/$ref_id/gene_exp.gmx";
+    push @outputs, [ "$outdir/gene_exp.gmx", 'diffexp_input_data' ] if -s "$outdir/gene_exp.gmx";
     push @outputs, [ $diffexp_file, 'job_result' ] if -s $diffexp_file;
     push @outputs, [ $diffexp_folder, 'folder' ] if -e $diffexp_folder and -d $diffexp_folder;
     
@@ -835,6 +850,13 @@ sub verify_cmd {
 sub save_output_files
 {
     my($app, $output) = @_;
+
+    my $diffexp_name = "diff_exp";
+    my $diffexp_folder = "$output/.$diffexp_name";
+    my $diffexp_file = "$output/$diffexp_name";
+
+    # copy over diffexp stuff
+    #$app->workspace->save_file_to_file("$diffexp_folder", {},"$output/.$diffexp_name", "job_result", 1);
     
     my %suffix_map = (
               txt => 'txt',
@@ -861,6 +883,8 @@ sub save_output_files
     {
         next if $p =~ /^\./;
         next if $p =~ /\.fna\z/;  
+        next if $p eq $diffexp_folder;
+        next if $p eq $diffexp_file;
         my @cmd = ("p3-cp", "-r", @suffix_map, "$output/$p", "ws:" . $app->result_folder);
         print "@cmd\n";
         my $ok = IPC::Run::run(\@cmd);
